+++ js/utils/check-string-length.js
@@ -1,3 +1,4 @@
 function checkStringLength (string, length) {
   return string.length <= length;
 }
+export{checkStringLength};
