+++ js/utils/getSum.js
@@ -1,0 +1,6 @@
+function getSum(aaa,bbb){
+  const description='Сумма двух чисел равна';
+  const sum=aaa+bbb;
+  return `${description  } ${ sum}`;
+}
+getSum(1,2);
